Update README to reflect changes
There were two changes that are now reflected in the README file.
The script now uses a socket lock to determine whether it is already
running and therefore it is no longer required to check the "pidof python"
in the crontab anymore.
It is now possible to (optionally) pass a custom config file as a
second commandline parameter.